Support OPTIONS request on LDPRms.
