+++ fcrepo-http-api/src/test/java/org/fcrepo/integration/http/api/AbstractVersioningIT.java
@@ -1,0 +1,211 @@
+/*
+ * Licensed to DuraSpace under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * DuraSpace licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except in
+ * compliance with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.fcrepo.integration.http.api;
+
+import static java.time.format.DateTimeFormatter.RFC_1123_DATE_TIME;
+import static javax.ws.rs.core.HttpHeaders.CONTENT_TYPE;
+import static javax.ws.rs.core.HttpHeaders.LINK;
+import static javax.ws.rs.core.HttpHeaders.LOCATION;
+import static javax.ws.rs.core.Response.Status.CREATED;
+import static javax.ws.rs.core.Response.Status.OK;
+import static org.apache.jena.graph.NodeFactory.createURI;
+import static org.apache.jena.rdf.model.ModelFactory.createDefaultModel;
+import static org.apache.jena.rdf.model.ResourceFactory.createProperty;
+import static org.fcrepo.http.api.FedoraVersioning.MEMENTO_DATETIME_HEADER;
+import static org.fcrepo.http.commons.domain.RDFMediaType.APPLICATION_LINK_FORMAT;
+import static org.fcrepo.kernel.api.FedoraTypes.FCR_VERSIONS;
+import static org.fcrepo.kernel.api.RdfLexicon.VERSIONED_RESOURCE;
+import static org.fcrepo.kernel.api.RdfLexicon.VERSIONING_TIMEMAP_TYPE;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
+import java.io.StringWriter;
+import java.time.LocalDateTime;
+import java.time.ZoneOffset;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+
+import javax.ws.rs.core.Link;
+
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.entity.StringEntity;
+import org.apache.http.util.EntityUtils;
+import org.apache.jena.graph.Node;
+import org.apache.jena.rdf.model.Model;
+import org.apache.jena.rdf.model.Property;
+import org.apache.jena.rdf.model.Resource;
+import org.apache.jena.riot.Lang;
+import org.apache.jena.riot.RDFDataMgr;
+import org.apache.jena.riot.RDFFormat;
+import org.apache.jena.riot.RDFLanguages;
+import org.junit.Before;
+
+import static org.fcrepo.kernel.api.RdfLexicon.CONSTRAINED_BY;
+import static org.fcrepo.kernel.api.RdfLexicon.MEMENTO_TYPE;
+
+/**
+ * @author lsitu
+ * @author bbpennel
+ */
+public abstract class AbstractVersioningIT extends AbstractResourceIT {
+
+    protected static final String VERSIONED_RESOURCE_LINK_HEADER = "<" + VERSIONED_RESOURCE.getURI()
+            + ">; rel=\"type\"";
+    protected static final String BINARY_CONTENT = "binary content";
+    protected static final String BINARY_UPDATED = "updated content";
+
+    protected static final String OCTET_STREAM_TYPE = "application/octet-stream";
+
+    protected static final Node MEMENTO_TYPE_NODE = createURI(MEMENTO_TYPE);
+    protected static final Node TEST_PROPERTY_NODE = createURI("info:test#label");
+
+    protected static final Property TEST_PROPERTY = createProperty("info:test#label");
+
+    final String MEMENTO_DATETIME =
+            RFC_1123_DATE_TIME.format(LocalDateTime.of(2000, 1, 1, 00, 00).atZone(ZoneOffset.UTC));
+
+    protected String subjectUri;
+    protected String id;
+
+    @Before
+    public void init() {
+        id = getRandomUniqueId();
+        subjectUri = serverAddress + id;
+    }
+
+    protected void verifyTimemapResponse(final String uri, final String id) throws Exception {
+        final List<Link> listLinks = new ArrayList<Link>();
+        listLinks.add(Link.fromUri(uri).rel("original").build());
+        listLinks.add(Link.fromUri(uri).rel("timegate").build());
+        listLinks
+                .add(Link.fromUri(uri + "/" + FCR_VERSIONS).rel("self").type(APPLICATION_LINK_FORMAT)
+                        .build());
+        final Link[] expectedLinks = listLinks.toArray(new Link[3]);
+
+        final HttpGet httpGet = getObjMethod(id + "/" + FCR_VERSIONS);
+        httpGet.setHeader("Accept", APPLICATION_LINK_FORMAT);
+        try (final CloseableHttpResponse response = execute(httpGet)) {
+            assertEquals("Didn't get a OK response!", OK.getStatusCode(), getStatus(response));
+            checkForLinkHeader(response, VERSIONING_TIMEMAP_TYPE, "type");
+            final List<String> bodyList = Arrays.asList(EntityUtils.toString(response.getEntity()).split(","));
+            final Link[] bodyLinks = bodyList.stream().map(String::trim).filter(t -> !t.isEmpty())
+                    .map(Link::valueOf).toArray(Link[]::new);
+            assertArrayEquals(expectedLinks, bodyLinks);
+        }
+    }
+
+    protected String createMemento(final String subjectUri, final String mementoDateTime, final String contentType,
+            final String body) throws Exception {
+        final HttpPost createVersionMethod = postObjMethod(id + "/" + FCR_VERSIONS);
+        if (contentType != null) {
+            createVersionMethod.addHeader(CONTENT_TYPE, contentType);
+        }
+        if (body != null) {
+            createVersionMethod.setEntity(new StringEntity(body));
+        }
+        if (mementoDateTime != null) {
+            createVersionMethod.addHeader(MEMENTO_DATETIME_HEADER, mementoDateTime);
+        }
+
+        // Create new memento of resource with updated body
+        try (final CloseableHttpResponse response = execute(createVersionMethod)) {
+            assertEquals("Didn't get a CREATED response!", CREATED.getStatusCode(), getStatus(response));
+            assertMementoDatetimeHeaderPresent(response);
+
+            return response.getFirstHeader(LOCATION).getValue();
+        }
+    }
+
+    protected String createContainerMementoWithBody(final String subjectUri, final String mementoDateTime)
+            throws Exception {
+
+        final String body = createContainerMementoBodyContent(subjectUri, "text/n3");
+        return createMemento(subjectUri, mementoDateTime, "text/n3", body);
+    }
+
+    protected String createContainerMementoBodyContent(final String subjectUri, final String contentType)
+        throws Exception {
+        // Produce new body from current body with changed triple
+        final String body;
+        final HttpGet httpGet = new HttpGet(subjectUri);
+        final Model model = createDefaultModel();
+        final Lang rdfLang = RDFLanguages.contentTypeToLang(contentType);
+        try (final CloseableHttpResponse response = execute(httpGet)) {
+            model.read(response.getEntity().getContent(), "", rdfLang.getName());
+        }
+        final Resource subjectResc = model.getResource(subjectUri);
+        subjectResc.removeAll(TEST_PROPERTY);
+        subjectResc.addLiteral(TEST_PROPERTY, "bar");
+
+        try (StringWriter stringOut = new StringWriter()) {
+            RDFDataMgr.write(stringOut, model, RDFFormat.NTRIPLES);
+            body = stringOut.toString();
+        }
+
+        return body;
+    }
+
+    protected void createVersionedContainer(final String id, final String subjectUri) throws Exception {
+        final HttpPost createMethod = postObjMethod();
+        createMethod.addHeader("Slug", id);
+        createMethod.addHeader(CONTENT_TYPE, "text/n3");
+        createMethod.addHeader(LINK, VERSIONED_RESOURCE_LINK_HEADER);
+        createMethod.setEntity(new StringEntity("<" + subjectUri + "> <info:test#label> \"foo\""));
+
+        try (final CloseableHttpResponse response = execute(createMethod)) {
+            assertEquals("Didn't get a CREATED response!", CREATED.getStatusCode(), getStatus(response));
+        }
+    }
+
+    protected void createVersionedBinary(final String id) throws Exception {
+        final HttpPost createMethod = postObjMethod();
+        createMethod.addHeader("Slug", id);
+        createMethod.addHeader(CONTENT_TYPE, OCTET_STREAM_TYPE);
+        createMethod.addHeader(LINK, VERSIONED_RESOURCE_LINK_HEADER);
+        createMethod.setEntity(new StringEntity(BINARY_CONTENT));
+
+        try (final CloseableHttpResponse response = execute(createMethod)) {
+            assertEquals("Didn't get a CREATED response!", CREATED.getStatusCode(), getStatus(response));
+        }
+    }
+
+    protected static void assertMementoDatetimeHeaderPresent(final CloseableHttpResponse response) {
+        assertNotNull("No memento datetime header set in response",
+                response.getHeaders(MEMENTO_DATETIME_HEADER));
+    }
+
+    protected static void assertMementoDatetimeHeaderMatches(final CloseableHttpResponse response,
+            final String expected) {
+        assertMementoDatetimeHeaderPresent(response);
+        assertEquals("Response memento datetime did not match expected value",
+                expected, response.getFirstHeader(MEMENTO_DATETIME_HEADER).getValue());
+    }
+
+    protected static void assertConstrainedByPresent(final CloseableHttpResponse response) {
+        final Collection<String> linkHeaders = getLinkHeaders(response);
+        assertTrue("Constrained by link header no present",
+                linkHeaders.stream().map(Link::valueOf)
+                        .anyMatch(l -> l.getRel().equals(CONSTRAINED_BY.getURI())));
+    }
+}
